+++ components/identity-mgt/org.wso2.carbon.identity.mgt.endpoint.util/src/main/java/org/wso2/carbon/identity/mgt/endpoint/util/client/model/recovery/v2/RecoveryInitRequest.java
@@ -16,68 +16,53 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.mgt.endpoint.util.client.model.passwordrecovery.v2;
+package org.wso2.carbon.identity.mgt.endpoint.util.client.model.recovery.v2;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.wso2.carbon.identity.mgt.endpoint.util.client.model.Property;
+import org.wso2.carbon.identity.mgt.endpoint.util.client.model.UserClaim;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
 /**
- * Request to recover username/password through given channel.
+ * Request to initiate an account recovery.
  **/
-public class RecoveryRequest {
+public class RecoveryInitRequest {
 
-    private String recoveryCode;
-    private String channelId;
+    private List<UserClaim> claims = null;
     private List<Property> properties = null;
 
     /**
-     * Recovery code for the user.
-     **/
-    public RecoveryRequest recoveryCode(String recoveryCode) {
+    * User claims to identify the user as UserClaim objects.
+    **/
+    public RecoveryInitRequest claims(List<UserClaim> claims) {
 
-        this.recoveryCode = recoveryCode;
+        this.claims = claims;
         return this;
     }
 
-    @JsonProperty("recoveryCode")
-    public String getRecoveryCode() {
+    @JsonProperty("claims")
+    public List<UserClaim> getClaims() {
 
-        return recoveryCode;
+        return claims;
+    }
+    public void setClaims(List<UserClaim> claims) {
+
+        this.claims = claims;
     }
 
-    public void setRecoveryCode(String recoveryCode) {
+    public RecoveryInitRequest addClaimsItem(UserClaim claimsItem) {
 
-        this.recoveryCode = recoveryCode;
-    }
-
-    /**
-     * Id of the notification channel that user prefers to get recovery notifications.
-     **/
-    public RecoveryRequest channelId(String channelId) {
-
-        this.channelId = channelId;
+        this.claims.add(claimsItem);
         return this;
     }
 
-    @JsonProperty("channelId")
-    public String getChannelId() {
-
-        return channelId;
-    }
-
-    public void setChannelId(String channelId) {
-
-        this.channelId = channelId;
-    }
-
     /**
-     * (OPTIONAL) Additional META properties.
-     **/
-    public RecoveryRequest properties(List<Property> properties) {
+    * (OPTIONAL) Additional META properties.
+    **/
+    public RecoveryInitRequest properties(List<Property> properties) {
 
         this.properties = properties;
         return this;
@@ -88,13 +73,12 @@ public class RecoveryRequest {
 
         return properties;
     }
-
     public void setProperties(List<Property> properties) {
 
         this.properties = properties;
     }
 
-    public RecoveryRequest addPropertiesItem(Property propertiesItem) {
+    public RecoveryInitRequest addPropertiesItem(Property propertiesItem) {
 
         if (this.properties == null) {
             this.properties = new ArrayList<>();
@@ -112,35 +96,33 @@ public class RecoveryRequest {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        RecoveryRequest recoveryRequest = (RecoveryRequest) o;
-        return Objects.equals(this.recoveryCode, recoveryRequest.recoveryCode) &&
-                Objects.equals(this.channelId, recoveryRequest.channelId) &&
-                Objects.equals(this.properties, recoveryRequest.properties);
+        RecoveryInitRequest initRequest = (RecoveryInitRequest) o;
+        return Objects.equals(this.claims, initRequest.claims) &&
+            Objects.equals(this.properties, initRequest.properties);
     }
 
     @Override
     public int hashCode() {
 
-        return Objects.hash(recoveryCode, channelId, properties);
+        return Objects.hash(claims, properties);
     }
 
     @Override
     public String toString() {
 
         StringBuilder sb = new StringBuilder();
-        sb.append("class RecoveryRequest {\n");
+        sb.append("class InitRequest {\n");
 
-        sb.append("    recoveryCode: ").append(toIndentedString(recoveryCode)).append("\n");
-        sb.append("    channelId: ").append(toIndentedString(channelId)).append("\n");
+        sb.append("    claims: ").append(toIndentedString(claims)).append("\n");
         sb.append("    properties: ").append(toIndentedString(properties)).append("\n");
         sb.append("}");
         return sb.toString();
     }
 
     /**
-     * Convert the given object to string with each line indented by 4 spaces
-     * (except the first line).
-     */
+    * Convert the given object to string with each line indented by 4 spaces
+    * (except the first line).
+    */
     private String toIndentedString(Object o) {
 
         if (o == null) {
